Add files via upload
ABSTRACT:
In structural analysis, the change of temperature in the truss member will lead it to either expand
or shrink. However, in statically indeterminate truss member, the length will stay the same due to
the support forces which is preventing it from expansion. Therefore, stress will develop in the
members due to changes in temperature. Similarly, fabrication errors will also produce additional
stress in the tresses.
In the finite element method, a discrete system is modeled with a finite number of nodes and a
finite number of elements. The behavior of each element is defined by stiffness or flexibility
relation. This relation can be further developed to establish a stiffness/flexibility matrix.
MATLAB program can be effectively used to model this system.
Objective:
1. To compute stress development in trusses and frames due to temperature and fabrication
changes.
2. Analyze plane truss supported by inclined support.
3. Compute member local stiffness matrix
4. Transform member stiffness matrix from local to the global coordinate system.
5. Compute the displacement along with the nodes
6. Compute the reaction in the plane accounting for fabrication errors and temperature
changes.